+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add organiaztion ID in batch order payload
 - Handle regex email and number input in B2B sale tunnel
 - Add locale context for components
+- Handle deletion of funding amount in B2B sale tunnel
 
 ## [3.2.1] - 2025-11-03